pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "spectra-ai"
version = "0.1.0"
description = "Runtime behavioral anomaly detection for AI agents"
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
authors = [
    { name = "Sushaan Kandukoori" },
]
keywords = [
    "ai",
    "agents",
    "anomaly-detection",
    "observability",
    "security",
    "monitoring",
    "opentelemetry",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: System :: Monitoring",
    "Topic :: Security",
    "Typing :: Typed",
]
dependencies = [
    "pydantic>=2.0,<3.0",
    "numpy>=1.24,<3.0",
    "opentelemetry-api>=1.20,<2.0",
    "opentelemetry-sdk>=1.20,<2.0",
    "httpx>=0.25,<1.0",
    "click>=8.0,<9.0",
]

[project.optional-dependencies]
dashboard = [
    "fastapi>=0.100,<1.0",
    "uvicorn>=0.23,<1.0",
]
ml = [
    "scikit-learn>=1.3,<2.0",
]
all = [
    "spectra-ai[dashboard,ml]",
]
dev = [
    "pytest>=7.0,<9.0",
    "pytest-asyncio>=0.21,<1.0",
    "pytest-cov>=4.0,<6.0",
    "ruff>=0.1,<1.0",
    "mypy>=1.5,<2.0",
]

[project.scripts]
spectra = "spectra.cli:main"

[project.urls]
Homepage = "https://github.com/sushaankandukoori/spectra"
Repository = "https://github.com/sushaankandukoori/spectra"
Issues = "https://github.com/sushaankandukoori/spectra/issues"
Documentation = "https://github.com/sushaankandukoori/spectra#readme"

[tool.hatch.build.targets.wheel]
packages = ["src/spectra"]

[tool.ruff]
target-version = "py311"
line-length = 88
src = ["src"]

[tool.ruff.lint]
select = [
    "E",    # pycodestyle errors
    "W",    # pycodestyle warnings
    "F",    # pyflakes
    "I",    # isort
    "N",    # pep8-naming
    "UP",   # pyupgrade
    "B",    # flake8-bugbear
    "SIM",  # flake8-simplify
    "TC",   # flake8-type-checking
    "RUF",  # ruff-specific
]
ignore = [
    "E501",     # line too long (handled by formatter)
    "TC001",    # runtime imports used in function bodies
    "TC002",    # third-party runtime imports
    "TC003",    # stdlib runtime imports
    "RUF007",   # pairwise preference is stylistic
    "B905",     # strict= on zip is not needed for padded slices
]

[tool.ruff.lint.isort]
known-first-party = ["spectra"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true

[[tool.mypy.overrides]]
module = "spectra.dashboard.app"
disallow_untyped_decorators = false

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
markers = [
    "integration: marks tests as integration tests",
]
